backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.15.1
celery==5.4.0
redis==5.0.5
rasterio==1.3.10
rio-cogeo==4.0.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.3
shapely==2.0.4
pyproj==3.6.1
reportlab==4.2.2
Pillow==12.1.1
python-multipart==0.0.22
httpx==0.27.0
pydantic-settings==2.3.1
aiofiles==23.2.1
python-jose[cryptography]==3.4.0
